requirements.txt
fastapi
uvicorn
python-multipart
pypdf
nltk
openai
python-dotenv
jinja2
pytest